Add endpoints (which do nothing yet)
