+++ contact-discovery-app/show_app_log.sh
