+++ contact-discovery-app/src/endpoints/all.ts
@@ -1,1 +1,1 @@
-export * from "./banking";
+export * from "./contact-discovery";
